+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Plugin will now detect and display if the server cannot be started (e.g. port is busy).
 - Support for multiple opened files and live logging cases.
+- Info line on the dashboard indicating how many files are currently loaded.
 
 ### Documentation
 - Minor formatting changes in README.md